refactor: 공통 usePressAnimation 훅 추가
