+++ packages/components/src/hooks/animations/index.ts
@@ -1,0 +1,1 @@
+export { usePressAnimation } from './usePressAnimation';
